% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_disprof.R
\name{f_disprof}
\alias{f_disprof}
\title{Dissimilarity profile analysis (DISPROF & SIMPROF)}
\usage{
f_disprof(Y, dis = "euc", iter = 1000, verb = TRUE, plt = FALSE, verb2 = TRUE)
}
\arguments{
\item{Y}{Matrix of response data (rows = observations, columns = variables)}

\item{dis}{Dissimilarity measure to be applied (see \code{vegan::vegdist}
  for additional details):
\itemize{
 \item Euclidean           (\code{'euc'})
 \item Bray Curtis         (\code{'bray'})
 \item Manhattan           (\code{'manhattan'})
 \item Gower's             (\code{'gower'})
 \item Alternative Gower's (\code{'altGower'})
 \item Canberra            (\code{'canberra'})
 \item Clark               (\code{'clark'})
 \item Kulczynski          (\code{'kulczynski'})
 \item Morisita            (\code{'morisita'})
 \item Horn                (\code{'horn'})
 \item Binomial            (\code{'binomial'})
 \item Cao                 (\code{'cao'})
}}

\item{iter}{Number of permutation for DISPROF analysis}

\item{verb}{Display results in the console window}

\item{plt}{Optionally create a DISPROF plot}

\item{verb2}{internal flag for f_mstCluster}
}
\value{
Returns a list structure that contains the following:
\itemize{
 \item \code{D}: observed dissimilarity profile
 \item \code{meanD}: mean permuted dissimilarity profife
 \item \code{Pi}: Pi statistic (= deviation of observed vs. permuted mean profile)
 \item \code{p}: randomized probability
}
}
\description{
This function follows Clarke et al.'s (2008) description of similarity profile
analysis (SIMPROF), but employs an equivalent approach based on dissimilarity
profile analysis (DISPROF).\cr
\cr
A null hypothesis of no multivariate structure is generated by calculating a
mean dissimilarity profile based on random permutations of the raw response
data. This approach follows a "Model 1: Environmental control over individual
species" permutational model described in Legendre & Legendre (2012:
page 619), whereby each column of data (i.e., the species) are permuted
separately.\cr
\cr
Note: when permuting the data, an internal check replaces NaN's with 1's;
this accounts for instances when permutation of the raw response data results
in more than one row consisting of all zeros, which in turn causes metrics
like the Bray-Curtis coefficient to returns NaN's after dividing 0 by 0.
}
\references{
Clarke, K. R., P. J. Somerfield, and R. N. Gorley. 2008.  Testing
  null hypotheses in exploratory community analyses: similarity profiles and
  biota-environmental linkage. J. Exp. Mar. Biol. Ecol. 366:56-69.
  (\href{https://doi.org/10.1016/j.jembe.2008.07.009}{Clarke et al., 2008})

Legendre, P. & L. Legendre. 2012. Numerical ecology. 3rd English
  ed. Elsevier Science BV, Amsterdam. [p. 619]
}
