% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_disprof_clust.R
\name{f_disprof_clust}
\alias{f_disprof_clust}
\title{Dissimilarity profile analysis (DISPROF) of a cluster analysis dendrogram}
\usage{
f_disprof_clust(
  Y,
  dis = "euc",
  txt = NULL,
  link = 1,
  mc = FALSE,
  iter = 1000,
  alpha = 0.05,
  tol = 0.005
)
}
\arguments{
\item{Y}{Matrix of response data (rows = observations, columns = variables)}

\item{dis}{Dissimilarity measure to be applied (see \code{vegan::vegdist}
  for additional details):
\itemize{
 \item Euclidean           (\code{'euc'})
 \item Bray Curtis         (\code{'bray'})
 \item Manhattan           (\code{'manhattan'})
 \item Gower's             (\code{'gower'})
 \item Alternative Gower's (\code{'altGower'})
 \item Canberra            (\code{'canberra'})
 \item Clark               (\code{'clark'})
 \item Kulczynski          (\code{'kulczynski'})
 \item Morisita            (\code{'morisita'})
 \item Horn                (\code{'horn'})
 \item Binomial            (\code{'binomial'})
 \item Cao                 (\code{'cao'})
}}

\item{txt}{Vector of row labels; generic labels are automatically generated if not provided}

\item{link}{Type of cluster analysis the be performed:
\enumerate{
 \item average    (UPGMA)
 \item centroid   (UPGMC)
 \item median     (WPGMC)
 \item single     (Shortest Distance)
 \item ward.D2    (Inner Squared Distance)
 \item mcquitty   (WPGMA)
}}

\item{mc}{Toggle progressive adjustment of p-value for multiple comparisons}

\item{iter}{Number of permutations for DISPROF analysis}

\item{alpha}{Significance level (default = 0.05)}

\item{tol}{Tolerance for rejecting a p-value > alpha (default = 0.005)}
}
\value{
Returns a list structure that contains the following:
\itemize{
 \item \code{grp}: vector specifying group membership for each row of Y
 \item \code{clusters}: explicit group membership
 \item \code{inc}: record of incremental change made to GRP
 \item \code{p}: corresponding randomized p-values
 \item \code{p_idx}: corresponding rows of Y used in the assessment
 \item \code{adj}: cell array of string indicating whether p-values were
 adjusted
 \item \code{Z}: cluster analysis tree linkages
 \item \code{cols}: vector indicating splits that were associated with
 significant p-values (=1) and those that were note (=0)
 \item \code{linkType}: method used for constructing cluster analysis tree
 \item \code{bin}: symmetric binary connectivity matrix defining pair-wise
 relationships as objects that are members of the same cluster (=FALSE) or
 members of different clusters (=TRUE)'
}
}
\description{
This function follows Clarke et al.'s (2008) description of similarity
profile analysis (SIMPROF), but employs an equivalent approach based on
dissimilarity profile analysis (DISPROF).\cr
\cr
This method is a form of agglomerative, hierarchical' cluster analysis.
}
\references{
Clarke, K. R., P. J. Somerfield, and R. N. Gorley. 2008.  Testing
  null hypotheses in exploratory community analyses: similarity profiles and
  biota-environmental linkage. J. Exp. Mar. Biol. Ecol. 366:56-69.
  (\href{https://doi.org/10.1016/j.jembe.2008.07.009}{Clarke et al., 2008})

Legendre, P. & L. Legendre. 2012. Numerical ecology. 3rd English
  ed. Elsevier Science BV, Amsterdam.
}
