% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_disprof_clustPlot.R
\name{f_disprof_clustPlot}
\alias{f_disprof_clustPlot}
\title{Plot dendrogram of a DISPROF-based custer analysis}
\usage{
f_disprof_clustPlot(result, orient = "V")
}
\arguments{
\item{result}{Output list from the 'f_disprof_clust' function}

\item{orient}{Plot orientation
\itemize{
 \item Vertical ('V')
 \item Horizontal ('H')
}}
}
\description{
This function plots the dendrogram based on the result from f_disprof_clust
analysis.
}
