% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dummy2cat.R
\name{f_dummy2cat}
\alias{f_dummy2cat}
\title{Dummy code to categorical variable}
\usage{
f_dummy2cat(X)
}
\arguments{
\item{X}{binary dummy codes created by f_dummy}
}
\value{
column vector of integers specifying factor levels or group
  membership
}
\description{
Convert dummy codes to a single categorical variable.
}
