% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_recode.R
\name{f_recode}
\alias{f_recode}
\title{Recode elements}
\usage{
f_recode(Y)
}
\arguments{
\item{Y}{vector or matrix of input variables (column-wise)}
}
\value{
vector or matrix recoded as consecutive integers
}
\description{
Recode elements of vector as consecutive integers. \cr
\cr
This function is used to convert treatment levels (e.g., "[2 2 2 5 5 5]") to
"[1 1 1 2 2 2]" so proper ANOVA design matrices can be constructed.
}
